+++ include/deal.II/distributed/p4est_wrappers.h
@@ -1,5 +1,20 @@
-#ifndef P4EST_WRAPPERS_H
-#define P4EST_WRAPPERS_H
+// ---------------------------------------------------------------------
+//
+// Copyright (C) 2016 - 2016 by the deal.II authors
+//
+// This file is part of the deal.II library.
+//
+// The deal.II library is free software; you can use it, redistribute
+// it, and/or modify it under the terms of the GNU Lesser General
+// Public License as published by the Free Software Foundation; either
+// version 2.1 of the License, or (at your option) any later version.
+// The full text of the license can be found in the file LICENSE at
+// the top level of the deal.II distribution.
+//
+// ---------------------------------------------------------------------
+
+#ifndef dealii__p4est_wrappers_h
+#define dealii__p4est_wrappers_h
 
 #include <deal.II/base/geometry_info.h>
 
@@ -531,4 +546,4 @@ DEAL_II_NAMESPACE_CLOSE
 
 #endif //DEAL_II_WITH_P4EST
 
-#endif //P4EST_WRAPPERS_H
+#endif //dealii__p4est_wrappers_h
