+++ source/distributed/p4est_wrappers.inst.in
@@ -1,6 +1,6 @@
 // ---------------------------------------------------------------------
 //
-// Copyright (C) 2010 - 2014 by the deal.II authors
+// Copyright (C) 2016 - 2016 by the deal.II authors
 //
 // This file is part of the deal.II library.
 //
