+++ source/fe/fe_tools_extrapolate.inst.in
@@ -1,0 +1,90 @@
+// ---------------------------------------------------------------------
+//
+// Copyright (C) 1998 - 2014 by the deal.II authors
+//
+// This file is part of the deal.II library.
+//
+// The deal.II library is free software; you can use it, redistribute
+// it, and/or modify it under the terms of the GNU Lesser General
+// Public License as published by the Free Software Foundation; either
+// version 2.1 of the License, or (at your option) any later version.
+// The full text of the license can be found in the file LICENSE at
+// the top level of the deal.II distribution.
+//
+// ---------------------------------------------------------------------
+
+
+
+for (deal_II_dimension : DIMENSIONS; deal_II_space_dimension :  SPACE_DIMENSIONS; VEC : SERIAL_VECTORS)
+{
+    namespace FETools
+    \{
+#if deal_II_dimension == deal_II_space_dimension
+    template
+    void extrapolate<deal_II_dimension>
+    (const DoFHandler<deal_II_dimension> &, const VEC &,
+     const DoFHandler<deal_II_dimension> &, VEC &);
+
+    template
+    void extrapolate<deal_II_dimension>
+    (const DoFHandler<deal_II_dimension> &, const VEC &,
+     const DoFHandler<deal_II_dimension> &, const ConstraintMatrix &,
+     VEC &);
+#endif
+    \}
+}
+
+
+
+for (deal_II_dimension : DIMENSIONS; deal_II_space_dimension :  SPACE_DIMENSIONS)
+{
+#ifdef DEAL_II_WITH_P4EST
+    namespace FETools
+    \{
+#if deal_II_dimension == deal_II_space_dimension
+
+#ifdef DEAL_II_WITH_PETSC
+    template
+    void extrapolate_parallel<deal_II_dimension, PETScWrappers::MPI::Vector, deal_II_space_dimension>
+    (const DoFHandler<deal_II_dimension,deal_II_space_dimension> &,
+     const PETScWrappers::MPI::Vector &,
+     const DoFHandler<deal_II_dimension,deal_II_space_dimension> &,
+     const ConstraintMatrix &,
+     PETScWrappers::MPI::Vector &);
+#endif
+
+#ifdef DEAL_II_WITH_TRILINOS
+    template
+    void extrapolate_parallel <deal_II_dimension, TrilinosWrappers::MPI::Vector, deal_II_space_dimension>
+    (const DoFHandler<deal_II_dimension,deal_II_space_dimension> &,
+     const TrilinosWrappers::MPI::Vector &,
+     const DoFHandler<deal_II_dimension,deal_II_space_dimension> &,
+     const ConstraintMatrix &,
+     TrilinosWrappers::MPI::Vector &);
+#endif
+
+#endif
+    \}
+#endif //DEAL_II_WITH_P4EST
+}
+
+
+
+for (deal_II_dimension : DIMENSIONS; deal_II_space_dimension :  SPACE_DIMENSIONS; scalar : REAL_SCALARS)
+{
+#ifdef DEAL_II_WITH_P4EST
+    namespace FETools
+    \{
+#if deal_II_dimension == deal_II_space_dimension
+    template
+    void extrapolate_parallel <deal_II_dimension, deal_II_space_dimension, scalar>
+    (const DoFHandler<deal_II_dimension,deal_II_space_dimension> &,
+     const LinearAlgebra::distributed::Vector<scalar> &,
+     const DoFHandler<deal_II_dimension,deal_II_space_dimension> &,
+     const ConstraintMatrix &,
+     LinearAlgebra::distributed::Vector<scalar> &);
+#endif
+    \}
+#endif //DEAL_II_WITH_P4EST
+}
+
